added data validation

+++ api/Controllers/CommentController.cs
@@ -11,53 +11,87 @@ namespace api.Controllers
 {
     [Route("api/comment")]
     [ApiController]
-    public class CommentController:ControllerBase
+    public class CommentController : ControllerBase
     {
         private readonly ICommentRepository _commentRepo;
         private readonly IStockRepository _stockRepo;
+
         public CommentController(ICommentRepository commentRepo, IStockRepository stockRepo)
         {
-            _commentRepo=commentRepo;
-            _stockRepo=stockRepo;
+            _commentRepo = commentRepo;
+            _stockRepo = stockRepo;
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll() { 
+        public async Task<IActionResult> GetAll()
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comments = await _commentRepo.GetAllAsync();
             var commentDto = comments.Select(c => c.ToCommentDto()).ToList();
             return Ok(commentDto);
         }
-        
+
         [HttpGet("{id:int}")]
-        
-        public async Task<IActionResult> GetById([FromRoute]int id){
-            var comment=await _commentRepo.GetByIdAsync(id);
-            if (comment==null)
+        public async Task<IActionResult> GetById([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var comment = await _commentRepo.GetByIdAsync(id);
+            if (comment == null)
             {
                 return NotFound();
             }
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
-        public async Task<IActionResult> Create([FromRoute]int stockId, CreateCommentDto commentDto)
+        [HttpPost("{stockId:int}")]
+        public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
-            if(!await _stockRepo.StockExists(stockId))
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            if (!await _stockRepo.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist");
             }
 
-            var commentModel=commentDto.ToCommentFromCreate(stockId);
+            var commentModel = commentDto.ToCommentFromCreate(stockId);
             await _commentRepo.CreateAsync(commentModel);
-            return CreatedAtAction(nameof(GetById),new { id = commentModel.Id }, commentModel.ToCommentDto());
+            return CreatedAtAction(
+                nameof(GetById),
+                new { id = commentModel.Id },
+                commentModel.ToCommentDto()
+            );
+        }
+
+        [HttpPut]
+        [Route("{id:int}")]
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var commentModel = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate(id));
+
+            if (commentModel == null)
+            {
+                return NotFound();
+            }
+            return Ok(commentModel.ToCommentDto());
         }
 
         [HttpDelete]
         [Route("{id:int}")]
-        public async Task<IActionResult> Delete([FromRoute]int id)
+        public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var commentModel = await _commentRepo.DeleteAsync(id);
-            if (commentModel==null)
+            if (commentModel == null)
             {
                 return NotFound("Comment does not exist");
             }
